Fleet Telemetry/requirements.txt
kafka-python==2.0.2
Faker==26.1.0